requirements.txt
autopep8
requests
pyyaml